Extend delete permissions for default user and enhance Google Analytics failure response
The default user permissions have been extended to include the `delete` action. For interaction with Google Analytics, the error handling has been improved. In case the service is unreachable, possibly due to an active ad-blocker, a suitable message is displayed, prompting to recheck the Analytics setup and rectify the issue. Furthermore, extended the functionality in the onHelpLinkClick method to handle cases where item.data is a function.