Implement months shift in shortcut-edit component and bump version
A months shift feature has been implemented in the shortcut-edit component for better period management. Additionally, the application version has been updated from '4.3.2' to '4.3.3'. Changes were also made in the locale messages, including the modification of the 'linenumber' fields to reflect the new lines of code after the additions and changes.