Add new features and modifications to settings, users, and environment
Introduced several enhancements including the addition of a mat-expansion-panel for URL settings in settings component, modification of the settings and view users components, and the addition of new demo and dev environment configurations. Additional files were added for environment configurations with corresponding asset replacements. Styles were also updated for the dialog actions. Updates to package.json to include commands for the demo build were also made.