Add 'pin' CSS class and enhance user validity check
Added a 'pin' class to the `settings.component.scss` improving the user interface and user experience. Enhanced the user validity check function on the 'settings.component.ts' by initializing a progress panel and setting a default empty info. Also, adjusted spaces in the markup in the 'gluc-adjust' component for better readability and consistency.