Update line numbers in locale messages file
This commit updates the line numbers for context in the messages.en-GB.xliff locale file. Line numbers have been incremented in various sections of the code to keep them accurate and up-to-date with the current state of the source files. Additionally, the URL example in the trans-unit id has been revised.