Add help icon to AAPS 3.0 data correction checkbox
The layout of AAPS 3.0 data correction checkbox has been modified to include a help icon next to it. Clicking on the help icon triggers a popup that provides relevant information about the checkbox. New CSS styles have been added to enhance the appearance of the help icon. Adjustments have also been made to title handling in the print template.