Update view-list component and whats-new HTML
The view-list component's expansionPanelClicked function has been correctly indented for better readability. In the Whats-New component HTML, changes have been made to the displayed information for version 4.3.9 and 2.0.12. Text has been added for version 4.3.9, and for version 2.0.12 part of the information has been replaced with a dynamic variable for flexibility in displaying messages.