Refactor PDF rendering logic to use Angular Renderer2.
Replaced direct DOM manipulation with Angular's `Renderer2` for improved security and maintainability in the PDF service. Updated logic to remove `app-root` and dynamically create an iframe for displaying the PDF in the same window. Added dependency injection for `RendererFactory2` and `DOCUMENT` to enable this change. This refactor aligns with Angular best practices and enhances compatibility with Angular's rendering engine.